requirements.txt
fastapi
uvicorn
sqlalchemy
pydentic
pandas
passlib[bcrypt]
python-jose